+++ src/app.js
@@ -9,6 +9,7 @@ import { list_saved_searches } from '../app/client/saved_searches/list_saved_sea
 import { load_event_listeners } from '../app/client/event_listeners/event_listeners.js'
 import { listener_authentication } from "../app/client/event_listeners/listener_authentication.js";
 import { check_authentication } from "../app/client/authentication/check_authentication.js";
+import { intro_modal } from "../app/client/intro/intro_modal.js";
 
 await check_authentication()
 listener_authentication()
@@ -21,6 +22,7 @@ listener_authentication()
 console.log('what is authenticated: '+ authenticated)
 if (authenticated)
 {
+    intro_modal()
     onload_searches()
     load_inspector('positive')
     load_inspector('negative')